Tronghieu

See merge request huydp_Irtech/eport!636

+++ eport-logistic/eport-logistic-portal/src/main/resources/static/js/logistic/receiveContEmpty/index.js
@@ -1,5 +1,5 @@
 var prefix = ctx + "logistic/receive-cont-empty";
-var interval, currentPercent;
+var interval, currentPercent, timeout;
 var dogrid = document.getElementById("container-grid"), hot;
 var shipmentSelected, shipmentDetails, shipmentDetailIds, sourceData, orderNumber = 0, currentVslNm;
 var contList = [], orders = [], processOrderIds;
@@ -725,6 +725,7 @@ function reloadShipmentDetail() {
     for (let i=0; i<checkList.length; i++) {
         $('#check'+i).prop('checked', false);
     }
+    $("#deleteBtn").prop("disabled", true);
     setLayoutRegisterStatus();
     loadShipmentDetail(shipmentSelected.id);
 }
@@ -967,25 +968,26 @@ function saveShipmentDetail() {
 
 // DELETE SHIPMENT DETAIL
 function deleteShipmentDetail() {
-    getDataSelectedFromTable(true);
-    $.modal.loading("Đang xử lý...");
-    $.ajax({
-        url: prefix + "/shipment/" + shipmentSelected.id + "/shipment-detail/" + shipmentDetailIds,
-        method: "delete",
-        success: function (result) {
-            if (result.code == 0) {
-                $.modal.alertSuccess(result.msg);
-                reloadShipmentDetail();
-            } else {
-                $.modal.alertError(result.msg);
-            }
-            $.modal.closeLoading();
-        },
-        error: function (result) {
-            $.modal.alertError("Có lỗi trong quá trình thêm dữ liệu, vui lòng liên hệ admin.");
-            $.modal.closeLoading();
-        },
-    });
+    if (getDataSelectedFromTable(true)) {
+        $.modal.loading("Đang xử lý...");
+        $.ajax({
+            url: prefix + "/shipment/" + shipmentSelected.id + "/shipment-detail/" + shipmentDetailIds,
+            method: "delete",
+            success: function (result) {
+                if (result.code == 0) {
+                    $.modal.alertSuccess(result.msg);
+                    reloadShipmentDetail();
+                } else {
+                    $.modal.alertError(result.msg);
+                }
+                $.modal.closeLoading();
+            },
+            error: function (result) {
+                $.modal.alertError("Có lỗi trong quá trình thêm dữ liệu, vui lòng liên hệ admin.");
+                $.modal.closeLoading();
+            },
+        });
+    }
 }
 
 // Handling logic
@@ -1075,7 +1077,7 @@ function finishVerifyForm(result) {
         connectToWebsocketServer();
 
         showProgress("Đang xử lý ...");
-        setTimeout(() => {
+        timeout = setTimeout(() => {
             setTimeout(() => {
                 hideProgress();
                 reloadShipmentDetail();
@@ -1108,6 +1110,8 @@ function onMessageReceived(payload) {
     let message = JSON.parse(payload.body);
     if (message.code != 0) {
 
+        clearTimeout(timeout);
+
         setProgressPercent(currentPercent=100);
         setTimeout(() => {
             hideProgress();
@@ -1125,6 +1129,8 @@ function onMessageReceived(payload) {
     } else {
         orderNumber--;
         if (orderNumber == 0) {
+            
+            clearTimeout(timeout);
 
             setProgressPercent(currentPercent=100);
             setTimeout(() => {
