+++ eport-logistic/eport-logistic-portal/src/main/resources/static/js/logistic/sendContEmpty/index.js
@@ -1,5 +1,5 @@
 var prefix = ctx + "logistic/send-cont-empty";
-var interval, currentPercent;
+var interval, currentPercent, timeout;
 var dogrid = document.getElementById("container-grid"), hot;
 var shipmentSelected, shipmentDetails, shipmentDetailIds, sourceData, processOrderIds;
 var contList = [];
@@ -651,6 +651,7 @@ function reloadShipmentDetail() {
     for (let i=0; i<checkList.length; i++) {
         $('#check'+i).prop('checked', false);
     }
+    $("#deleteBtn").prop("disabled", true);
     setLayoutRegisterStatus();
     loadShipmentDetail(shipmentSelected.id);
 }
@@ -873,25 +874,26 @@ function saveShipmentDetail() {
 
 // DELETE SHIPMENT DETAIL
 function deleteShipmentDetail() {
-    getDataSelectedFromTable(true);
-    $.modal.loading("Đang xử lý...");
-    $.ajax({
-        url: prefix + "/shipment/" + shipmentSelected.id + "/shipment-detail/" + shipmentDetailIds,
-        method: "delete",
-        success: function (result) {
-            if (result.code == 0) {
-                $.modal.alertSuccess(result.msg);
-                reloadShipmentDetail();
-            } else {
-                $.modal.alertError(result.msg);
-            }
-            $.modal.closeLoading();
-        },
-        error: function (result) {
-            $.modal.alertError("Có lỗi trong quá trình thêm dữ liệu, vui lòng liên hệ admin.");
-            $.modal.closeLoading();
-        },
-    });
+    if (getDataSelectedFromTable(true)) {
+        $.modal.loading("Đang xử lý...");
+        $.ajax({
+            url: prefix + "/shipment/" + shipmentSelected.id + "/shipment-detail/" + shipmentDetailIds,
+            method: "delete",
+            success: function (result) {
+                if (result.code == 0) {
+                    $.modal.alertSuccess(result.msg);
+                    reloadShipmentDetail();
+                } else {
+                    $.modal.alertError(result.msg);
+                }
+                $.modal.closeLoading();
+            },
+            error: function (result) {
+                $.modal.alertError("Có lỗi trong quá trình thêm dữ liệu, vui lòng liên hệ admin.");
+                $.modal.closeLoading();
+            },
+        });
+    }
 }
 
 // Handling logic
@@ -980,7 +982,7 @@ function finishVerifyForm(result) {
         connectToWebsocketServer();
 
         showProgress("Đang xử lý ...");
-        setTimeout(() => {
+        timeout = setTimeout(() => {
             setTimeout(() => {
                 hideProgress();
                 reloadShipmentDetail();
@@ -1015,6 +1017,7 @@ function onError(error) {
 }
 
 function onMessageReceived(payload) {
+    clearTimeout(timeout);
     setProgressPercent(currentPercent=100);
     setTimeout(() => {
         hideProgress();
